~Organization +FF 4chan thread styles
I reorganized all of the code, enabling for more convenient viewing when
the chrome extension code is added later, and the tampermonkey code is
added after that.

Additionally, I removed the list of posts that a particular 4chan post
is replied to in (>>49509470 >>495791847095 >>49574896789, etc) and
replaced it with a button showing the number of times it has been
replied to; This button will also eventually produce a dropdown of all
of the posts.