+FF 4chan thread styles ~README
Added Bootstrap 3.2.0 JS and JQuery 1.11.1 scripts to bottom of <head>,
which has absolutely enfuriated s.4cdn.org/js/extension.878.js which
logs lots of errors to the console and powers image expansion and stuff.

Anyways, a button that drops down to the list of links to posts that
reply to the current post was added. Thread stats were re-worded to be
clearer. Removed some superfluous console logging that was commented
out.

Also expanded upon the README.