~FF 4chan thread styling (gallery mode)
Made image gallery actually load the  first time you switch to that
view. Initial load speed of the page increases, if you don't want to use
the gallery mode then your resources aren't expended on it, and the
information in gallery mode is more correct.
Updated todo list in readme.